+++ CHANGELOG.md
@@ -23,12 +23,18 @@
 * `SBBIconTextButton`: Added field `focusNode`
 * `SBBModalSheet`: Added field `useRootNavigator`
 * `SBBModalSheet`: Added constructor `custom` for header customizing
+* `SBBToast`: Added field `bottom`
 ### Changed
 * `SBBHeader`: Renamed field `onPressedSBBSignet` to `onPressedLogo`
 * `SBBHeader`: Renamed field `sbbSignetTooltip` to `logoTooltip`
 * `SBBOnboarding`: Padding now defined by `SBBThemeData.defaultRootContainerPadding` 
+* `SBBTextStyles`: Adjusted `fontSize` and `height` values to match the current specifications
+* `SBBListHeader`: Adjusted paddings to match the current specifications
+* `SBBTextField`: Adjusted paddings to match the current specifications
 ### Deprecated
 ### Removed
+* `SBBListHeader`: Removed fields `icon` and `onCallToAction` to match the current specifications
+* `SBBTextField`: Removed field `alignLabelWithHint`
 ### Fixed
 * `SBBHeader`: SBB Logo now excluded from semantics if `onPressedLogo` is `null`
 * `SBBModalSheet`: Fine line that sometimes was visible below the header is now gone